refactor: move icons
