+++ components/MarketParticipant/index.tsx
@@ -3,10 +3,10 @@ import { CSSProperties, FC, useEffect, useState } from 'react';
 import { classNames } from '@/utils/index';
 import { fadeInDown } from '@/utils/animations';
 import { RoleId } from '@/types/roles';
-import HammerIcon from '../walkthroughs/icons/HammerIcon';
-import PoundcashTag from '../walkthroughs/icons/PoundcashTag';
-import CartPlus from '../walkthroughs/icons/CartPlus';
-import OfferIcon from '../walkthroughs/icons/OfferIcon';
+import HammerIcon from '../../icons/HammerIcon';
+import PoundcashTag from '../../icons/PoundcashTag';
+import CartPlus from '../../icons/CartPlus';
+import OfferIcon from '../../icons/OfferIcon';
 import { MarketParticipantTitle } from '../MarketParticipantTitle';
 import { Products } from '../../types/products';
 import { BiodiversityCount } from '../BiodiversityCount';
