+++ components/ProjectDetails/index.tsx
@@ -1,8 +1,8 @@
 import { ChangeEvent, FC, FormEvent, useRef, useState } from 'react';
 import { motion } from 'framer-motion';
 import { fadeIn } from '@/utils/animations';
-import BiodiversityIconGray from '@/components/walkthroughs/icons/BiodiversityIcon';
-import NutrientsIcon from '@/components/walkthroughs/icons/NutrientsIcon';
+import BiodiversityIconGray from 'icons/BiodiversityIcon';
+import NutrientsIcon from 'icons/NutrientsIcon';
 import { classNames } from '@/utils/index';
 import { RoleId } from '@/types/roles';
 import { roles } from '../../data/roles';
