+++ icons/BiodiversityIcon.tsx
