+++ icons/MarketBackgroundLeft.tsx
