+++ icons/OfferIcon.tsx
